Move SecuredropOwner to directory app
